Remove delete and modification methods, add addEvent method
