+++ portfolio/src/main/java/com/google/sps/data/CalendarEventsGroup.java
@@ -7,13 +7,15 @@ import java.util.Comparator;
 import java.util.List;
 
 /** This class models a group of calendar events. */
-public class CalendarGroup {
+public class CalendarEventsGroup {
+
+  // This variable that represents a group of calendar events will
+  // always be sorted in ascending order based on their start time.
   public List<CalendarEvent> events;
-  public List<TimeRange> allFreeTimeRanges;
 
   // These two variables represent the start and end of
   // possible schedule time. When these times are passed in
-  // to instantiate a CalendarGroup class, these times are
+  // to instantiate a this class, these times are
   // already adjusted based on user's working hour times
   // and the current timestamp of when the user starts
   // the scheduling process.
@@ -31,16 +33,16 @@ public class CalendarGroup {
    *     working hour start time and the current timestamp.
    * @param endTime: end of possible scheduling blocks, of type Instant.
    */
-  public CalendarGroup(
+  public CalendarEventsGroup(
       List<CalendarEvent> events, Instant overallStartTime, Instant overallEndTime) {
     if (events == null) {
-      throw new IllegalArgumentException("Calendar Group needs a collection of events");
+      throw new IllegalArgumentException("Events cannot be null");
     }
     if (overallStartTime == null) {
-      throw new IllegalArgumentException("Calendar Group needs the start of scheduling hours");
+      throw new IllegalArgumentException("Overall start of scheduling hours cannot be null");
     }
     if (overallEndTime == null) {
-      throw new IllegalArgumentException("Calendar Group needs the end of scheduling hours");
+      throw new IllegalArgumentException("Overall end of scheduling hours cannot be null");
     }
     if (overallStartTime.isAfter(overallEndTime)) {
       throw new IllegalArgumentException("Start time cannot be after end time");
@@ -48,7 +50,6 @@ public class CalendarGroup {
     this.events = events;
     this.overallStartTime = overallStartTime;
     this.overallEndTime = overallEndTime;
-    this.allFreeTimeRanges = calculateFreeTimeRanges();
   }
 
   /**
@@ -56,7 +57,7 @@ public class CalendarGroup {
    * lie completely inside the possible scheduling hours. Originally written by tomasalvarez,
    * modified by hollyyuqizheng.
    */
-  public List<TimeRange> calculateFreeTimeRanges() {
+  public List<TimeRange> getFreeTimeRanges() {
     Collections.sort(events, sortByEventStartTimeAscending);
 
     List<TimeRange> possibleTimes = new ArrayList<TimeRange>();
@@ -85,42 +86,13 @@ public class CalendarGroup {
     return possibleTimes;
   }
 
-  /** Getter for the list of all free time ranges. */
-  public List<TimeRange> getFreeTimeRanges() {
-    return allFreeTimeRanges;
-  }
-
   /**
-   * Deletes a time range from the list of all free times.
-   *
-   * @return the modifed list of free time ranges.
+   * Adds a new event into the event group. Returns the new collection of events, sorted based on
+   * start time.
    */
-  public List<TimeRange> deleteFreeTimeRange(TimeRange timeRange) {
-    if (allFreeTimeRanges.contains(timeRange)) {
-      allFreeTimeRanges.remove(timeRange);
-    } else {
-      throw new IllegalArgumentException("This time range does not represent any free time range.");
-    }
-    return allFreeTimeRanges;
-  }
-
-  /**
-   * Modifies an original time range with a new one.
-   *
-   * @return the modified list of free times.
-   */
-  public List<TimeRange> modifyFreeTimeRange(TimeRange originalTimeRange, TimeRange newTimeRange) {
-    if (originalTimeRange.start().isAfter(newTimeRange.start())
-        || originalTimeRange.end().isBefore(newTimeRange.end())) {
-      throw new IllegalArgumentException(
-          "New time range can only be shorter than original time range");
-    }
-
-    allFreeTimeRanges.remove(originalTimeRange);
-    allFreeTimeRanges.add(newTimeRange);
-
-    // Sorts all free time ranges before returning.
-    Collections.sort(allFreeTimeRanges, TimeRange.sortByTimeRangeStartTimeAscending);
-    return allFreeTimeRanges;
+  public List<CalendarEvent> addNewEvent(CalendarEvent event) {
+    events.add(event);
+    Collections.sort(events, sortByEventStartTimeAscending);
+    return events;
   }
 }
